backend/requirements.txt
Flask
Flask-CORS
requests
google-generativeai
python-dotenv
